remove src/lib/zod/index
co-locate api handler input schemas with handlers

move input schema validation helpers to api-lib/requests/schema; more refactoring to come
