+++ api-lib/signature.ts
@@ -12,11 +12,11 @@ import { ethers } from 'ethers';
 import { SiweMessage } from 'siwe';
 import { z } from 'zod';
 
-import { composeCrossClientAuthRequestBody } from '../src/lib/zod';
 import { zEthAddressOnly } from '../src/lib/zod/formHelpers';
 
 import { INFURA_PROJECT_ID } from './config';
 import { errorLog } from './HttpError';
+import { composeCrossClientAuthRequestBody } from './requests/schema';
 
 const PERSONAL_SIGN_REGEX = /0x[0-9a-f]{130}/;
