+++ api-test/helpers/epochs.ts
@@ -1,8 +1,40 @@
 import { DateTime } from 'luxon';
+import { z } from 'zod';
 
-import { createEpochInput } from '../../src/lib/zod';
+import { zStringISODateUTC } from '../../src/lib/zod/formHelpers';
 
 import type { GQLClientType } from './common';
+
+const createEpochInput = z
+  .object({
+    circle_id: z.number().int().positive(),
+    start_date: zStringISODateUTC,
+    description: z.string().min(10).max(100).optional(),
+    repeat: z.number().int().min(0).max(3),
+    days: z
+      .number()
+      .min(1, 'Must be at least one day.')
+      .max(100, 'cant be more than 100 days'),
+    grant: z.number().positive().min(1).max(1000000000).optional(),
+  })
+  .strict()
+  .superRefine((val, ctx) => {
+    let message;
+    if (val.days > 7 && val.repeat === 1) {
+      message =
+        'You cannot have more than 7 days length for a weekly repeating epoch.';
+    } else if (val.days > 28 && val.repeat === 2) {
+      message =
+        'You cannot have more than 28 days length for a monthly repeating epoch.';
+    }
+
+    if (message) {
+      ctx.addIssue({
+        code: z.ZodIssueCode.custom,
+        message,
+      });
+    }
+  });
 
 type EpochInput = typeof createEpochInput['_type'];
