+++ api/hasura/actions/_handlers/endEpoch.ts
@@ -5,7 +5,7 @@ import { z } from 'zod';
 import { authCircleAdminMiddleware } from '../../../../api-lib/circleAdmin';
 import { adminClient } from '../../../../api-lib/gql/adminClient';
 import { errorResponseWithStatusCode } from '../../../../api-lib/HttpError';
-import { composeHasuraActionRequestBody } from '../../../../src/lib/zod';
+import { composeHasuraActionRequestBody } from '../../../../api-lib/requests/schema';
 import { endEpochHandler } from '../../cron/epochs';
 
 Settings.defaultZone = 'utc';
