+++ api/hasura/actions/_handlers/logoutUser.ts
@@ -1,8 +1,8 @@
 import type { VercelRequest, VercelResponse } from '@vercel/node';
 
 import { adminClient } from '../../../../api-lib/gql/adminClient';
+import { HasuraUserSessionVariables } from '../../../../api-lib/requests/schema';
 import { verifyHasuraRequestMiddleware } from '../../../../api-lib/validate';
-import { HasuraUserSessionVariables } from '../../../../src/lib/zod';
 
 async function handler(req: VercelRequest, res: VercelResponse) {
   const { hasuraProfileId } = HasuraUserSessionVariables.parse(
@@ -17,14 +17,10 @@ async function handler(req: VercelRequest, res: VercelResponse) {
             tokenable_type: { _eq: 'App\\Models\\Profile' },
           },
         },
-        {
-          affected_rows: true,
-        },
+        { affected_rows: true },
       ],
     },
-    {
-      operationName: 'logoutUser_deleteToken',
-    }
+    { operationName: 'logoutUser_deleteToken' }
   );
   return res.status(200).json({ id: hasuraProfileId });
 }
