+++ api/hasura/actions/_handlers/updateEpoch.ts
@@ -5,8 +5,8 @@ import { z } from 'zod';
 import { authCircleAdminMiddleware } from '../../../../api-lib/circleAdmin';
 import { adminClient } from '../../../../api-lib/gql/adminClient';
 import { errorResponseWithStatusCode } from '../../../../api-lib/HttpError';
+import { composeHasuraActionRequestBody } from '../../../../api-lib/requests/schema';
 import { Awaited } from '../../../../api-lib/ts4.5shim';
-import { composeHasuraActionRequestBody } from '../../../../src/lib/zod';
 
 import {
   zEpochInputParams,
